Get all the indices from Opensearch eligible for reindexing
